requirements.txt
huggingface_hub
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
llama-cpp-python